move exception_details_invalid back to exception_details
